Tests/LibWeb/Text/expected/quirks-mode-case-insensitive-class-selector.txt

ParentNode.querySelector matches class selectors case-insensitively in quirks mode: true
